+++ pkg/deployment/reconcile/plan_builder_high.go
@@ -59,6 +59,7 @@ func (r *Reconciler) createHighPlan(ctx context.Context, apiObject k8sutil.APIOb
 		ApplyIfEmpty(r.createRebalancerCheckPlan).
 		ApplyIfEmpty(r.createMemberFailedRestoreHighPlan).
 		ApplyWithBackOff(BackOffCheck, time.Minute, r.emptyPlanBuilder)).
+		ApplyIfEmptyWithBackOff(TimezoneCheck, time.Minute, r.createTimezoneUpdatePlan).
 		Apply(r.createBackupInProgressConditionPlan). // Discover backups always
 		Apply(r.createMaintenanceConditionPlan).      // Discover maintenance always
 		Apply(r.cleanupConditions)                    // Cleanup Conditions