+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (Logging) Internal client trace
 - (QA) Member maintenance feature
 - (Feature) Extract Pod Details
+- (Feature) Add Timezone management
 
 ## [1.2.15](https://github.com/arangodb/kube-arangodb/tree/1.2.15) (2022-07-20)
 - (Bugfix) Ensure pod names not too long